move ray scheduling ut to test_ray_scheduling.py
